--- a/src/.init.lua
+++ b/src/.init.lua
@@ -1,9 +1,20 @@
+sqlite3 = require "lsqlite3"
+
 function IntToIPString(n)
     return string.format("%d.%d.%d.%d",
         (n >> 24) & 255,
         (n >> 16) & 255,
         (n >> 8) & 255,
         (n >> 0) & 255)
+end
+
+function ConnectDb()
+    if not db then
+        db = sqlite3.open("db.sqlite3")
+        db:busy_timeout(1000)
+        db:exec[[PRAGMA journal_mode=WAL]]
+        db:exec[[PRAGMA synchronous=NORMAL]]
+    end
 end
 
 -- Redbean's global route handler
@@ -21,14 +32,10 @@ function OnHttpRequest()
     end
 end
 
--- Setup() isn't called by the server,
--- you just call it yourself in the
--- redbean REPL
-function Setup()
-    local sqlite3 = require "lsqlite3"
-    local db = sqlite3.open("db.sqlite3", sqlite3.OPEN_READWRITE + sqlite3.OPEN_CREATE)
+function OnServerStart()
+    ConnectDb()
     local res = db:exec[[
-        CREATE TABLE "land" (
+        CREATE TABLE IF NOT EXISTS "land" (
             ip INTEGER PRIMARY KEY,
             nick TEXT NOT NULL CHECK (length(nick) = 8),
             created_at TEXT DEFAULT (STRFTIME('%Y-%m-%d %H:%M:%f', 'NOW'))
@@ -39,4 +46,6 @@ function Setup()
     else
         print(db:errmsg())
     end
+    db:close()
+    db = nil
 end