--- a/src/claim.lua
+++ b/src/claim.lua
@@ -13,9 +13,7 @@ elseif re.search([[^[a-zA-Z0-9]{8}$]], name) ~= name then
     SetStatus(400)
     Write("Name must be ASCII alphanumeric")
 else
-    local sqlite3 = require "lsqlite3"
-    local db = sqlite3.open("db.sqlite3", sqlite3.OPEN_READWRITE)
-    db:exec[[PRAGMA writable_schema=ON]] -- until redbean supports strict
+    ConnectDb()
     local stmt = db:prepare([[
         INSERT INTO land (ip, nick) VALUES (?1, ?2)
         ON CONFLICT (ip) DO UPDATE SET (nick) = (?2)