--- a/src/summary.lua
+++ b/src/summary.lua
@@ -29,9 +29,7 @@ else
     biggest = 0xFFFFFFFF
 end
 
-local sqlite3 = require "lsqlite3"
-local db = sqlite3.open("db.sqlite3", sqlite3.OPEN_READONLY)
-db:exec[[PRAGMA writable_schema=ON]] -- until redbean supports strict
+ConnectDb()
 local stmt = db:prepare([[
     SELECT nick, COUNT(ip) as score FROM land WHERE ip >= ?1 AND ip <= ?2 GROUP BY nick;
 ]])