This change helps TurfWar handle 420k qps of read traffic.

```
main jart@nightmare:~/vendor/turfwar$ wrk -t 12 -c 120 http://10.10.10.124:8080/
Running 10s test @ http://10.10.10.124:8080/
  12 threads and 120 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   380.99us    5.13ms 210.87ms   99.72%
    Req/Sec    35.45k    20.32k  118.37k    68.99%
  4253968 requests in 10.10s, 11.74GB read
Requests/sec: 421152.43
Transfer/sec:      1.16GB
```